Grid: move current selection up/down when scrolling
- Make the selection work with lines in lines_above and lines_below
- Todo: update selection end when scrolling with mouse button being held
- Todo: figure out how to move selection when new characters are added.

+++ zellij-utils/src/input/mouse.rs
@@ -11,21 +11,21 @@ pub struct Position {
 impl Position {
     pub fn new(line: u16, column: u16) -> Self {
         Self {
-            line: Line(line as usize),
+            line: Line(line as isize),
             column: Column(column as usize),
         }
     }
 
     pub fn relative_to(&self, position_and_size: &PositionAndSize) -> Self {
         Self {
-            line: Line(self.line.0 - position_and_size.y),
+            line: Line(self.line.0 - position_and_size.y as isize),
             column: Column(self.column.0 - position_and_size.x),
         }
     }
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Deserialize, Serialize, PartialOrd)]
-pub struct Line(pub usize);
+pub struct Line(pub isize);
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Deserialize, Serialize, PartialOrd)]
 pub struct Column(pub usize);
